# Optimize `std::regex` in `IRLoggerParser` C++ class