--- a/src/nqm/irimager/irlogger_parser.cpp
+++ b/src/nqm/irimager/irlogger_parser.cpp
@@ -23,7 +23,7 @@ static const std::unordered_map<std::string, LogLevel>
  * @throws std::logic_error if parsing the IRLogger log line failed.
  */
 static std::tuple<LogLevel, std::string> parse_line(const std::string &line) {
-  auto regex =
+  const thread_local auto regex =
       std::regex(R"""((\w+)\ \[([\w.:\d]*)\]\ \@\ ([\d\.]+s)\ :(.*))""",
                  std::regex_constants::ECMAScript);